src/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
# slack-sdk>=3.20.2